Extends ReactEcharts class
