--- packages/react-ui/src/custom-components/echarts-for-react/helper/is-equal.js
@@ -1,6 +1,0 @@
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.isEqual = void 0;
-var tslib_1 = require("tslib");
-var fast_deep_equal_1 = (0, tslib_1.__importDefault)(require("fast-deep-equal"));
-exports.isEqual = fast_deep_equal_1.default;
-//# sourceMappingURL=is-equal.js.map
