--- packages/react-ui/src/custom-components/echarts-for-react/helper/is-function.js
@@ -1,7 +1,0 @@
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.isFunction = void 0;
-function isFunction(v) {
-    return typeof v === 'function';
-}
-exports.isFunction = isFunction;
-//# sourceMappingURL=is-function.js.map
