--- packages/react-ui/src/custom-components/echarts-for-react/helper/is-string.js
@@ -1,7 +1,0 @@
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.isString = void 0;
-function isString(v) {
-    return typeof v === 'string';
-}
-exports.isString = isString;
-//# sourceMappingURL=is-string.js.map
