--- packages/react-ui/src/custom-components/echarts-for-react/types.js
@@ -1,2 +1,0 @@
-Object.defineProperty(exports, "__esModule", { value: true });
-//# sourceMappingURL=types.js.map
